requirements.txt
streamlit==1.37.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

transformers==4.30.2
tokenizers==0.13.3
sentence-transformers==2.2.2
huggingface-hub==0.19.4

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

PyPDF2==3.0.1
docx2txt==0.8

fpdf2==2.7.8
Pillow==10.3.0

seaborn==0.13.2
matplotlib==3.8.0
plotly==5.22.0

tqdm==4.66.1
requests==2.31.0
regex==2023.10.3
filelock==3.12.2
packaging==23.2
safetensors==0.4.1
